# Faster `beam_search` sampling