--- a/whisper.cpp
+++ b/whisper.cpp
@@ -24,7 +24,6 @@
 #include <regex>
 #include <random>
 #include <unordered_set>
-#include <iostream>
 
 #if defined(_MSC_VER)
 #pragma warning(disable: 4244 4267) // possible loss of data