**PR (v1):** The KV cache update logic in the Master is quite simple and heavy-handed. First, a temporary cache is established, and all the KV cache is stored in this temporary cache. After the beam search is completed, the KV cache is updated based on the results of the beam search, pulling from the temporary cache to replace the old KV cache. If we assume the `beam_size` is `X`, then with the above method we need to read and write the KV cache `2X` times, which is very inefficient. In reality, we often don't need to update the decoder's KV cache (zero-copy), and even if we do need to update, we only need to update a portion and can directly copy without going through the temporary cache (one-copy). Lastly, we can also detect if there are read-write conflicts to reduce the number of times we write to the temporary cache and then copy (two-copy).

**PR (v2):** The calculation of `whisper_sample_token_topk` in Master is highly inefficient. First, it uses `logits_id.clear()` to clear all elements in the vector, and then it uses a loop to extract the logit value from logits. Along with the current index, a new `std::pair` is created and then pushed to the end of `logits_id` using `push_back`. Because the length of logits is determined by `n_vocab`, and `n_vocab` in whisper is `51,865`, each time we calculate `whisper_sample_token_topk` using the method in Master, we need to create (via `logits_id.push_back`) ~~and destroy (via `logits_id.clear`)~~ `51,865` `std::pair` objects, which is highly inefficient. In fact, we can use `logits_id.resize()` to limit the size of `logits_id`. Since the new size is less than or equal to the old capacity, there will be no memory allocation, and we can directly assign values to the old pairs using the index, thus avoiding creation and destruction.

**PR(v3):** Adopted ggerganov's suggestion to make ·kv_bufs· static, avoiding memory allocation and deallocation, thereby significantly improving performance.

**PR(v4):** Added the pointer swapping function, which slightly improved performance.

**PR(v5):** Updated the logic for determining `two-copy`, reducing unnecessary memory copying, slightly improving performance.

This PR optimizes both the KV cache update logic and the calculation of `whisper_sample_token_topk`, significantly reducing memory read-write operations, thereby saving a lot of time. 

The average sample time has decreased from **`1.56 ms/run`** to **`0.69 ms/run`**, reducing the computational time by approximately **`~55%`**. @ggerganov 

**Master:** `i7-12700H` `diffusion2023-07-03.wav` `ggml-model-whisper-base.bin`
| `-bs 1` `-bo 1` | `-bs 2` `-bo 2` | `-bs 3` `-bo 3` | `-bs 4` `-bo 4` | `-bs 5` `-bo 5` |
| :-----: | :----: | :----: | :----: | :----: |
| 0.57 ms/run | 1.56 ms/run | 1.55 ms/run | 1.58 ms/run | 1.55 ms/run |

<details> <summary>Older Versions</summary>

**This PR (v1):** `i7-12700H` `diffusion2023-07-03.wav` `ggml-model-whisper-base.bin`
| `-bs 1` `-bo 1` | `-bs 2` `-bo 2` | `-bs 3` `-bo 3` | `-bs 4` `-bo 4` | `-bs 5` `-bo 5` |
| :-----: | :----: | :----: | :----: | :----: |
| 0.57 ms/run | 0.96 ms/run | 1.01 ms/run | 1.07 ms/run | 1.14 ms/run |

**This PR (v2):** `i7-12700H` `diffusion2023-07-03.wav` `ggml-model-whisper-base.bin`
| `-bs 1` `-bo 1` | `-bs 2` `-bo 2` | `-bs 3` `-bo 3` | `-bs 4` `-bo 4` | `-bs 5` `-bo 5` |
| :-----: | :----: | :----: | :----: | :----: |
| 0.57 ms/run | 0.76 ms/run | 0.81 ms/run | 0.87 ms/run | 0.94 ms/run |

**This PR (v3):** `i7-12700H` `diffusion2023-07-03.wav` `ggml-model-whisper-base.bin`
| `-bs 1` `-bo 1` | `-bs 2` `-bo 2` | `-bs 3` `-bo 3` | `-bs 4` `-bo 4` | `-bs 5` `-bo 5` |
| :-----: | :----: | :----: | :----: | :----: |
| 0.57 ms/run | 0.69 ms/run | 0.72 ms/run | 0.74 ms/run | 0.77 ms/run |

**This PR (v4):** `i7-12700H` `diffusion2023-07-03.wav` `ggml-model-whisper-base.bin`
| `-bs 1` `-bo 1` | `-bs 2` `-bo 2` | `-bs 3` `-bo 3` | `-bs 4` `-bo 4` | `-bs 5` `-bo 5` |
| :-----: | :----: | :----: | :----: | :----: |
| 0.57 ms/run | 0.67 ms/run | 0.70 ms/run | 0.72 ms/run | 0.75 ms/run |

</details>

**This PR (v5):** `i7-12700H` `diffusion2023-07-03.wav` `ggml-model-whisper-base.bin`
| `-bs 1` `-bo 1` | `-bs 2` `-bo 2` | `-bs 3` `-bo 3` | `-bs 4` `-bo 4` | `-bs 5` `-bo 5` |
| :-----: | :----: | :----: | :----: | :----: |
| 0.57 ms/run | 0.64 ms/run | 0.68 ms/run | 0.70 ms/run | 0.73 ms/run |