@ggerganov 2023-09-05
> First, it uses logits_id.clear() to clear all elements in the vector, and then it uses a loop to extract the logit value from logits.

This is not slower than the new approach, because all that `clear()` does is change the internal elements counter to 0. The memory of the vector remains allocated. But anyway, the new implementation is also fine