@bobqianic 2023-09-07
I've now visualized the logic for updating the KV cache across different versions, as shown in the figure. I've made the changes in the new version based on the suggestions you provided earlier. If there aren't any bugs, it's ready to be merged. Thanks. @ggerganov 
### Master:
<img width="662" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/ae31de44-a40a-486c-b7f4-54d2436466b3">

### PR(v1):
<img width="666" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/2fb0f212-4b85-4126-9a61-d428d4eefd93">

### PR(v4):
<img width="663" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/211db6f4-ac5b-49a1-95a4-393c77bf3c10">