@ggerganov 2023-09-07
The test program that you made is very useful. You should try to add it as a unit test in `tests`

It might be a good idea to create a wiki page with your idea and plots from this PR. Or at the very least, you can just link to this PR. But the goal is this information to remain available and easy to find in the future.

You can also add a wiki entry for your previous PR about matching OpenAI input. 
I've sent you a collaborator invite